+++ src/test/java/org/gridsuite/merge/orchestrator/server/MergeOrchestratorControllerTest.java
@@ -28,7 +28,6 @@ import java.nio.charset.StandardCharsets;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -37,8 +36,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.http.MediaType.APPLICATION_OCTET_STREAM;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 /**
@@ -63,9 +61,6 @@ public class MergeOrchestratorControllerTest extends AbstractEmbeddedCassandraSe
     @Inject
     IgmRepository igmRepository;
 
-    @Inject
-    ProcessConfigRepository processConfigRepository;
-
     @MockBean
     private IgmQualityCheckService igmQualityCheckService;
 
@@ -87,30 +82,6 @@ public class MergeOrchestratorControllerTest extends AbstractEmbeddedCassandraSe
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
-        ArrayList<String> tsos = new ArrayList<>();
-        tsos.add("FR");
-        tsos.add("ES");
-        tsos.add("PT");
-        processConfigRepository.save(new ProcessConfigEntity("SWE", tsos, false));
-    }
-
-    @Test
-    public void configTest() throws Exception {
-        mvc.perform(get("/" + VERSION + "/configs")
-                .contentType(APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON))
-                .andExpect(content().json("[{\"process\":\"SWE\",\"tsos\":[\"FR\",\"ES\",\"PT\"]}]"));
-
-        mvc.perform(get("/" + VERSION + "/configs/SWE")
-                .contentType(APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON))
-                .andExpect(content().json("{\"process\":\"SWE\",\"tsos\":[\"FR\",\"ES\",\"PT\"]}"));
-
-        mvc.perform(delete("/" + VERSION + "/configs/SWE")
-                .contentType(APPLICATION_JSON))
-                .andExpect(status().isOk());
     }
 
     @Test
